restructure: remove tiny_chunk funcs to separate file
